Handle duplicate times
If we know a fraction of a second, enter that in column D and the results will show the full seconds still, but the ranking will take the fraction into account.

So Phil Rayner: 23:44.67; Lawrence Cox: 23:44.82 - Phil is the faster.

Changes made to both CC and Ev10 workbooks.

Also - League tables still had FC in a column header.  Changed to DC.  Thanks to Andrew Weatherby for spotting this error.