Table corrections
Per Jon Welsford:

In the seniors sheet all is well as are the cox league pages.  However the veteran, women, road bikes*2, juveniles, juniors sheets totals for number of events and total points don’t seem to include the results for last Tuesday.  The points are in the appropriate column but not included in the totals.