Championship scoring changes
This year, there is a requirement to have competed in at least two events which are not 10mile TTs in order to be eligible.

This means that the championship tables (Seniors, Vets, Women's, etc.) will show a rider's championship score as 0 until they meet this requirement.

Note that non-qualifying riders - i.e. those who have competed in fewer than 2 non-ten events - are ranked according to the score for their best 9 ten-mile times and their (at most one) non-ten-mile event.