Competitor import fix
When a new competitor row is being dealt with, don't consider existing rows with a later creation date so that the new row can slot in before it.

This is for a real scenario such as a late update of the membership list after an event has taken place, but with a "simulated_import_date" of the day before the event.