Introduces three separate Rank tracking properties
AllEventsRank - based on the score for all events competed in - no limits

TenMileCompetitionRank - based on the score for the best n ten-mile events that the competitor has ridden.  In 2025, n was 8, but that can vary depending on the total nuber of 10-mile events in the calendar.

FullCompetitionRank - based on the score for the best n events of any type, but mandating that the rider competes in at least two non-ten-mile events before they become eligible.  In 2025, this was best two non-tens and then the best 9 remaining events of any distance.  Total = 11 events.